com.vgis.python.rs/vgis_rs/requirements.txt
GDAL==3.4.3
Pillow==9.3.0
pyproj==3.6.1